rename 'image' module to 'integration'; this is consistent with jxl-oxide naming
